I figured out how to use multiple inheritance, which was probably how I should have done it the first time...
